# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "CONFESC-API",
    "description": "",
    "license": {
      "name": ""
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/inscrever": {
      "post": {
        "tags": [
          "Inscrito"
        ],
        "operationId": "post_inscrito",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InscritoWithCursosDTO"
              }
            }
          },
          "required": true
        },
        "responses": {}
      }
    },
...
    "/inscrito/{id}": {
      "get": {
        "tags": [
          "Inscrito"
        ],
        "operationId": "get_inscrito_by_id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {}
      }
    },
    "/inscrito/{id}/cursos": {
      "get": {
        "tags": [
          "Inscrito"
        ],
        "operationId": "get_inscrito_cursos",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {}
      }
    },
    "/noticia": {
      "get": {
        "tags": [
          "Noticia"
        ],
        "operationId": "get_noticias",
        "responses": {}
      },
      "post": {
        "tags": [
          "Noticia"
        ],
        "operationId": "post_noticia",
        "requestBody": {
          "description": "Noticia do store in the database",
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/NewNoticia"
              }
            }
          },
          "required": true
        },
        "responses": {}
      }
    },
...
    "/noticia/{id}": {
      "get": {
        "tags": [
          "Noticia"
        ],
        "operationId": "get_noticia_by_id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {}
      },
      "put": {
        "tags": [
          "Noticia"
        ],
        "operationId": "put_noticia",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/NewNoticia"
              }
            }
          },
          "required": true
        },
        "responses": {}
      },
      "delete": {
        "tags": [
          "Noticia"
        ],
        "operationId": "delete_noticia",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {}
      }
    },
    "/noticia/{id}/imagem": {
      "get": {
        "tags": [
          "Noticia"
        ],
        "operationId": "get_noticia_imagem",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {}
      }
    }
  },
  "components": {
    "schemas": {
...
      "InscritoWithCursosDTO": {
        "type": "object",
        "required": [
          "nome",
          "nome_cracha",
          "email",
          "cpf",
          "modalidade_nome",
          "modalidade_preco",
          "instituicao"
        ],
        "properties": {
          "cpf": {
            "type": "string"
          },
          "cursos": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int32"
            },
            "nullable": true
          },
          "email": {
            "type": "string"
          },
          "instituicao": {
            "type": "string"
          },
          "modalidade_nome": {
            "type": "string"
          },
          "modalidade_preco": {
            "type": "integer",
            "format": "int32"
          },
          "nome": {
            "type": "string"
          },
          "nome_cracha": {
            "type": "string"
          }
        }
      },
...
      "NewNoticia": {
        "type": "object",
        "required": [
          "titulo",
          "conteudo",
          "autor",
          "imagem"
        ],
        "properties": {
          "autor": {
            "type": "string"
          },
          "conteudo": {
            "type": "string"
          },
          "imagem": {
            "type": "string",
            "format": "binary"
          },
          "titulo": {
            "type": "string"
          }
        }
      }
    }
  }
}